requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
sqlparse==0.5.3
